backend/suggestions/requirements.txt
Django==5.1.1
djangorestframework==3.15.2
django-rest-framework==0.1.0
openai==1.44.1
py-eureka-client==0.11.10